scripts/requirements.txt
feedparser==6.0.11
supabase==2.10.0
python-dotenv==1.0.1